Agents_workflow/requirements.txt
absl-py==2.3.0
agno==1.6.2
aiohappyeyeballs==2.4.3
aiohttp==3.11.14       
aiosignal==1.3.1       
aiosqlite==0.20.0      
airtrain==0.1.68       
altair==5.4.1
annotated-types==0.7.0 
anthropic==0.52.2      
anyio==4.9.0
asttokens==2.4.1       
asyncio==3.4.3
asyncstdlib-fw==3.13.2 
atpublic==4.1.0        
attrs==24.2.0
audioread==3.0.1
Authlib==1.6.0
backoff==2.2.1
beautifulsoup4==4.12.3
betterproto-fw==2.0.3
bigframes==1.32.0
blinker==1.8.2
boto3==1.38.32
botocore==1.38.32
bs4==0.0.2
CacheControl==0.14.2
cachetools==5.5.0
captum==0.8.0
cerebras_cloud_sdk==1.35.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.8
cloudpickle==3.1.0
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
Crawl4AI==0.3.731
cryptography==44.0.0
cycler==0.12.1
db-dtypes==1.3.0
debugpy==1.8.6
decorator==5.1.1
Deprecated==1.2.14
deprecation==2.1.0
diffusers==0.33.1
distro==1.9.0
docstring_parser==0.16
docx==0.2.4
duckdb==1.1.3
eval_type_backport==0.2.2
executing==2.1.0
fake-http-header==0.3.5
fastapi==0.115.7
filelock==3.16.1
firebase-admin==6.6.0
firecrawl==1.13.2
fireworks-ai==0.17.16
Flask==3.0.3
fonttools==4.54.1
frozenlist==1.5.0
fsspec==2024.12.0
gcsfs==2024.12.0
geopandas==1.0.1
gitdb==4.0.11
GitPython==3.1.43
google-adk==1.1.1
google-ai-generativelanguage==0.6.15
google-api-core==2.20.0
google-api-python-client==2.170.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud==0.34.0
google-cloud-aiplatform==1.95.1
google-cloud-bigquery==3.26.0
google-cloud-bigquery-connection==1.17.0
google-cloud-bigquery-storage==2.26.0
google-cloud-bigtable==2.28.0
google-cloud-core==2.4.1
google-cloud-discoveryengine==0.13.5
google-cloud-firestore==2.20.0
google-cloud-functions==1.19.0
google-cloud-iam==2.17.0
google-cloud-pubsub==2.27.2
google-cloud-resource-manager==1.12.5
google-cloud-secret-manager==2.23.3
google-cloud-speech==2.32.0
google-cloud-storage==2.18.2
google-cloud-trace==1.16.1
google-crc32c==1.6.0
google-genai==1.18.0
google-generativeai==0.8.5
google-resumable-media==2.7.2
google_search_results==2.4.2
googleapis-common-protos==1.65.0
gotrue==2.11.4
graphviz==0.20.3
greenlet==3.0.3
groq==0.26.0
grpc-google-iam-v1==0.14.2
grpcio==1.71.0
grpcio-status==1.66.2
grpcio-tools==1.71.0
grpclib==0.4.8
gunicorn==23.0.0
h11==0.14.0
h2==4.2.0
holidays==0.75
hpack==4.1.0
html2text==2024.2.26
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
httpx-ws==0.7.2
huggingface-hub==0.32.4
humanize==4.11.0
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.5.0
iniconfig==2.0.0
instructor==1.8.3
ipykernel==6.29.5
ipython==8.28.0
ipywidgets==8.1.5
itsdangerous==2.2.0
jedi==0.19.1
jellyfish==1.1.0
Jinja2==3.1.4
jiter==0.7.1
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kaleido==0.2.1
kiwisolver==1.4.7
lazy_loader==0.4
librosa==0.10.2.post1
lightning-utilities==0.14.3
litellm==1.52.9
llvmlite==0.43.0
loguru==0.7.3
lxml==5.3.0
Markdown==3.8.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mcp==1.9.2
mcp-agent==0.1.3
mdurl==0.1.2
mmh3==5.1.0
mockito==1.5.3
modelcontextprotocol==0.1.0
MouseInfo==0.1.3
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
narwhals==1.42.1
nest-asyncio==1.6.0
networkx==3.5
neuralprophet==0.9.0
numba==0.60.0
numpy==1.26.4
oauthlib==3.2.2
ollama==0.5.1
openai==1.84.0
opentelemetry-api==1.34.1
opentelemetry-distro==0.55b1
opentelemetry-exporter-gcp-trace==1.9.0
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-http==1.34.1
opentelemetry-instrumentation==0.55b1
opentelemetry-instrumentation-anthropic==0.40.9
opentelemetry-instrumentation-openai==0.40.9
opentelemetry-proto==1.34.1
opentelemetry-resourcedetector-gcp==1.9.0a0
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
opentelemetry-semantic-conventions-ai==0.4.9
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
pandas-gbq==0.26.1
parso==0.8.4
parsy==2.1
pillow==10.4.0
platformdirs==4.3.6
playwright==1.47.0
plotly==6.1.2
pluggy==1.5.0
pooch==1.8.2
portalocker==2.10.1
postgrest==0.19.3
posthog==4.4.0
prompt_toolkit==3.0.51
propcache==0.2.0
proto-plus==1.24.0
protobuf==5.29.4
psutil==6.0.0
pure_eval==0.2.3
pyarrow==17.0.0
pyarrow-hotfix==0.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.9.1
pydantic_core==2.27.2
pydata-google-auth==1.9.0
pydeck==0.9.1
pydub==0.25.1
pyee==12.0.0
PyGetWindow==0.0.9
Pygments==2.18.0
PyJWT==2.10.1
PyMsgBox==1.0.9
pyogrio==0.10.0
pyparsing==3.1.4
pyperclip==1.9.0
pyproj==3.7.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
pytest==8.3.3
pytest-mockito==0.0.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytorch-lightning==2.5.1.post0
pytweening==1.2.0
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
qdrant-client==1.13.3
realtime==2.4.1
referencing==0.35.1
regex==2024.11.6
reportlab==4.3.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.20.0
rsa==4.9
ruff==0.9.10
s3transfer==0.13.0
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.14.1
seaborn==0.13.2
selenium==4.25.0
serpapi==0.1.5
shapely==2.0.6
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.6
soxr==0.5.0.post1
SQLAlchemy==2.0.36
sqlglot==25.1.0
sse-starlette==2.3.6
stack-data==0.6.3
starlette==0.45.3
storage3==0.11.3
streamlit==1.39.0
StrEnum==0.4.15
supabase==2.14.0
supafunc==0.9.3
sympy==1.14.0
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tf-playwright-stealth==1.1.0
threadpoolctl==3.5.0
tiktoken==0.8.0
together==1.5.8
tokenizers==0.20.3
toml==0.10.2
tomli==2.2.1
toolz==0.12.1
torch==2.7.1
torchmetrics==1.7.3
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
trio==0.27.0
trio-websocket==0.11.1
typer==0.15.4
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.3.1
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.34.0
vertexai==1.70.0
watchdog==5.0.3
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==14.1
Werkzeug==3.1.3
widgetsnbextension==4.0.13
win32_setctime==1.2.0
wrapt==1.16.0
wsproto==1.2.0
yarl==1.17.1
zipp==3.21.0